Update: turns out i'm not crazy.
Found the code for the model in MatLab + C (C has been easier to understand)
A lot was misreported in the supplementary document.
Currently diving through and making the corrections for the channels which were not predicting well - BK (corrected in this update)
need to do Ical and TCal next.
> Experimental temperature is different - correct for
> parameter values are different.
>parameters in paper are in different units to code - correct>